Finished #166716486 Finished restructuring the folder structure
