+++ api/v1/controllers/index.js
@@ -3,7 +3,7 @@ const routes = require('express').Router();
 const properties = require('./properties');
 const users = require('./users')
 
-routes.use('/properties', properties);
-routes.use('/users', users);
+routes.use('/v1/properties', properties);
+routes.use('/v1/users', users);
 
 module.exports = routes;
